requirements.txt
Django==2.0.1
django-admin-interface==0.5.9
django-colorfield==0.1.14
django-flat-responsive==2.0
django-flat-theme==1.1.4
django-jet==1.0.6
django-material==1.2.0
djangorestframework-jwt==1.11.0
gunicorn==19.7.1
PyJWT==1.5.3
pytz==2017.3
